+++ src/app/services/sites.service.ts
@@ -30,6 +30,22 @@ export class SitesService {
                 catchError(this.handleError<any>('getEventSites', []))
             );
     }
+
+    public getFilteredSites(urlParams: string): Observable<Site[]> {
+        console.log('URL Parameters passed: ' + urlParams);
+        return this.httpClient
+            .get(APP_SETTINGS.SITES_URL + '/FilteredSites.json?' + urlParams)
+            .pipe(
+                tap((response) => {
+                    console.log(
+                        'getFilteredSites response received: ' +
+                            JSON.stringify(response)
+                    );
+                    return response;
+                }),
+                catchError(this.handleError<any>('getFilteredSites', []))
+            );
+    }
     /* istanbul ignore next */
     private handleError<T>(operation = 'operation', result?: T) {
         return (error: any): Observable<T> => {