+++ CHANGELOG.md
@@ -9,6 +9,21 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+-   Added EventType and EventState Filters to filter the Events choice list
+-   Added site filter options for: HWMs Surveyed, HWM Only, Sensor Only, Surveyed RP Only, Pre-Deployed Bracket, RDG Only, Has OP Defined
+
+### Changed
+
+-
+
+### Fixed
+
+-
+
+## [v0.3.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.3.0) - 2020-12-08
+
+### Added
+
 -   Added prerequisite for at least one layer to be checked before the legend appears
 -   Added popup message for when a layer is removed due to user zooming out
 -   Added min zoom warning in layer control
@@ -18,8 +33,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 -   Adjusted color of second navbar
 -   Changed "HWMS Surveyed" from checkbox to radio buttons
-
-### Fixed
 
 ## [v0.2.0](https://github.com/USGS-WiM/stnweb2/releases/tag/v0.2.0) - 2020-11-24